# api/openapi/gateway.yaml
openapi: 3.1.0
info:
  title: Aizorix Public API
  version: "1.0.0"
  description: >
    Public REST surface exposed by the API gateway. Internal service-to-service
    traffic uses gRPC (see api/proto). Most paths are generated from proto
    google.api.http annotations; REST-only surfaces (Stripe webhooks, file
    uploads) are defined here explicitly. All endpoints require a Bearer JWT
    unless marked `security: []`.
servers:
  - url: https://api.aizorix.com
  - url: http://localhost:8080
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Project:
      type: object
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        description: { type: string }
        budget_type: { type: string, enum: [fixed, hourly] }
        budget_min_cents: { type: integer, format: int64 }
        budget_max_cents: { type: integer, format: int64 }
        currency: { type: string }
        status: { type: string }
        skills: { type: array, items: { type: string } }
paths:
...
  /v1/projects:
    get:
      summary: Search/browse published projects (delegates to search service)
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: skills, in: query, schema: { type: array, items: { type: string } } }
        - { name: budget_type, in: query, schema: { type: string, enum: [fixed, hourly] } }
        - { name: min_budget, in: query, schema: { type: integer } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Project' } }
                  next_cursor: { type: string }
    post:
      summary: Create a project (client only)
      responses: { "201": { description: Created } }
  /v1/projects/{id}:
    get:
      summary: Get a project
      parameters: [{ name: id, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { "200": { description: OK }, "404": { description: Not found } }
  /v1/proposals:
    get:
      summary: List proposals for a project (client) — flat surface
      parameters:
        - { name: project_id, in: query, required: true, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { type: string } }
      responses: { "200": { description: OK } }
    post:
      summary: Submit a proposal (freelancer only); project_id is in the body
      responses: { "201": { description: Created }, "402": { description: Insufficient connects }, "409": { description: Duplicate proposal } }
...